refactor(acp): extract the nori-acp-host crate (#529)
Extract the agent-agnostic ACP hosting machinery from nori-acp into a new
Layer-0 crate, nori-acp-host:

- connection/: agent subprocess spawning, the AcpConnection wire client, and
  MCP server wiring
- translator.rs: the generic bridge between ACP wire types and
  codex-protocol events
- registry.rs: agent registry and distribution resolution
- patch.rs: file-change and diff helpers
- error_category.rs: AcpErrorCategory and categorize_acp_error, split out of
  backend/mod.rs

enhanced_error_message stays in the nori-acp backend — it is user-facing UX
wording, not a wire concern.

nori-acp re-exports the moved modules, so downstream consumers are unchanged
here; the TUI moves onto direct crate deps in a follow-up. This is the crate
an external project would pull in to host any ACP agent from its own
frontend.

Building the crate standalone caught two dependency-feature gaps masked by
workspace feature unification (tokio-util "compat" and
agent-client-protocol-schema "unstable") — exactly the class of bug the
extraction exists to surface.
