+++ nori-rs/README.md
@@ -16,7 +16,9 @@ progressively adopted or removed (see `docs/specs/crate-layering.md`).
 |-------|---------|
 | `cli/` (`nori-cli`) | The shipped `nori` binary: dispatch and subcommands |
 | `tui/` (`nori-tui`) | Ratatui interactive terminal interface |
-| `acp/` (`nori-acp`) | ACP backend: agent registry, connection, session runtime |
+| `acp/` (`nori-acp`) | ACP backend session harness: session runtime, transcripts, hooks |
+| `acp-host/` (`nori-acp-host`) | Agent-agnostic ACP hosting: subprocess spawn, wire client, registry |
+| `nori-config/` | Nori config layer (`~/.nori/cli/config.toml`) |
 | `nori-protocol/` | Session-runtime types over the ACP schema |
 | `sandbox/` (`codex-sandbox`) | Sandboxed exec engine: Seatbelt, Landlock/seccomp, Windows restricted tokens |
 | `installed/` (`nori-installed`) | Install detection and analytics |
