+++ nori-rs/acp-host/Cargo.toml
@@ -1,0 +1,49 @@
+[package]
+edition = "2024"
+name = "nori-acp-host"
+version = { workspace = true }
+
+[lib]
+doctest = false
+name = "nori_acp_host"
+path = "src/lib.rs"
+
+[lints]
+workspace = true
+
+[dependencies]
+agent-client-protocol = { workspace = true }
+agent-client-protocol-schema = { workspace = true, features = ["unstable"] }
+anyhow = { workspace = true }
+codex-protocol = { workspace = true }
+base64 = { workspace = true }
+codex-rmcp-client = { workspace = true }
+codex-utils-string = { workspace = true }
+diffy = { workspace = true }
+futures = { workspace = true }
+libc = { workspace = true }
+nori-config = { workspace = true }
+oauth2 = "5"
+regex = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+tokio = { workspace = true, features = [
+    "io-util",
+    "macros",
+    "process",
+    "rt-multi-thread",
+    "sync",
+    "time",
+] }
+tokio-util = { workspace = true, features = ["compat"] }
+tracing = { workspace = true, features = ["log"] }
+which = { workspace = true }
+
+[dev-dependencies]
+pretty_assertions = { workspace = true }
+rmcp = { workspace = true, features = [
+    "client",
+    "transport-streamable-http-client-reqwest",
+] }
+serial_test = { workspace = true }
+tempfile = { workspace = true }
