+++ nori-rs/docs.md
@@ -11,7 +11,7 @@ This is the Rust implementation of Nori, a terminal-based AI coding assistant. T
 The `nori-rs` directory is the root of a Cargo workspace containing all Rust code for the project. The workspace is organized into focused crates that handle specific concerns:
 
 - **Entry points**: `tui/` provides the main TUI application, `cli/` provides the `nori` binary dispatch and sandbox debug utilities
-- **ACP integration**: `acp/` handles communication with ACP-compliant agents spawned as local subprocesses, and owns session state, transcripts, Nori config, and the moved-in leaf helpers (shell parsing, notifications, custom prompts, compact constants)
+- **ACP integration**: `acp/` (`nori-acp`) is the session harness -- it owns the ACP backend session runtime, transcripts, hooks, goals, and moved-in leaf helpers (shell parsing, notifications, custom prompts, compact constants). The agent-agnostic hosting machinery (subprocess spawning, wire client, registry, translator) lives in `acp-host/` (`nori-acp-host`) and the Nori config layer in `nori-config/`; both are re-exported through `nori-acp` so consumer paths are unchanged
 - **Shared infrastructure**: `core/` contains configuration, authentication, and model/provider metadata consumed by the frontends (not by `acp/`)
 - **Protocol definitions**: `protocol/`, `app-server-protocol/`, `mcp-types/` define shared type vocabularies
 - **Sandboxing**: `sandbox/` (`codex-sandbox`) owns the sandboxed exec engine and platform sandbox selection; `linux-sandbox/`, `windows-sandbox-rs/`, `execpolicy/` provide the platform-specific pieces
@@ -19,7 +19,7 @@ The `nori-rs` directory is the root of a Cargo workspace containing all Rust cod
 
 Most shared crates still follow the inherited `codex-` prefix convention (for example `codex-core` and `codex-protocol`), while Nori-owned entrypoint crates now use `nori-` names such as `nori-acp`, `nori-protocol`, `nori-installed`, and `nori-tui`.
 
-The workspace is converging on a three-layer structure -- publishable ACP-host leaves at the bottom, a headless session harness in the middle, thin frontends on top -- per `@/docs/specs/crate-layering.md`. Milestones already landed: dead Codex-engine subsystems deleted from `codex-core`, the `nori-tui` `nori-config` cargo feature removed (Nori config is the only path), protocol types imported directly from `codex-protocol` (core's re-export detour deleted), `nori-acp`'s dependency on `codex-core` fully severed, and the sandboxed-execution engine extracted from `codex-core` into `codex-sandbox` (`@/nori-rs/sandbox/`).
+The workspace is converging on a three-layer structure -- publishable ACP-host leaves at the bottom, a headless session harness in the middle, thin frontends on top -- per `@/docs/specs/crate-layering.md`. Milestones already landed: dead Codex-engine subsystems deleted from `codex-core`, the `nori-tui` `nori-config` cargo feature removed (Nori config is the only path), protocol types imported directly from `codex-protocol` (core's re-export detour deleted), `nori-acp`'s dependency on `codex-core` fully severed, the sandboxed-execution engine extracted from `codex-core` into `codex-sandbox` (`@/nori-rs/sandbox/`), the config layer extracted into `nori-config` (`@/nori-rs/nori-config/`), and the agent-agnostic ACP hosting machinery extracted into the Layer-0 `nori-acp-host` crate (`@/nori-rs/acp-host/`).
 
 ### Core Implementation
 
@@ -27,7 +27,9 @@ The TUI drives user interaction through a Ratatui-based interface. When using AC
 
 Architecture:
 - nori-tui (TUI) -> Terminal User Interface
-  - nori-acp -> ACP Agent Connection -> External ACP Agents (claude, etc); depends only on codex-protocol among inherited crates
+  - nori-acp -> ACP session harness -> External ACP Agents (claude, etc); depends only on codex-protocol among inherited crates
+    - nori-acp-host -> agent-agnostic ACP hosting leaf (subprocess spawn, wire client, registry, translator); re-exported through nori-acp
+    - nori-config -> Nori config layer (~/.nori/cli/config.toml); re-exported through nori-acp as `config`
   - codex-core -> Config/Auth infrastructure for the frontends (nori-tui, nori-cli)
   - codex-sandbox -> Sandboxed exec engine and platform sandbox selection; imported directly by core, tui, cli, and linux-sandbox
   - codex-protocol -> Shared type vocabulary, imported directly by every consumer
